+++ meridian/model/eda/constants.py
@@ -16,7 +16,8 @@
 
 # EDA Engine constants
 COST_PER_MEDIA_UNIT = 'cost_per_media_unit'
-
+RSQUARED_GEO = 'rsquared_geo'
+RSQUARED_TIME = 'rsquared_time'
 VARIABLE_1 = 'var1'
 VARIABLE_2 = 'var2'
 CORRELATION = 'correlation'