+++ CHANGELOG.md
@@ -23,6 +23,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+* Add EDA check for treatment/control geo and time collinearity.
+
 ## [1.4.0] - 2025-12-08
 
 * Introduce modules needed for Meridian Scenario Planner and add